podpora pro kombinovani upgradu do jednoho vytvoreny zakladni upgrady bugfixy runspeed typ zmenen z int na float

+++ Assets/scripts/AI/MonsterAI.cs
@@ -635,7 +635,7 @@ namespace Assets.scripts.AI
 			currentAction = null;
 		}
 
-        protected virtual IEnumerator MoveAction(Vector3 target, bool fixedRotation, int fixedSpeed=-1)
+        protected virtual IEnumerator MoveAction(Vector3 target, bool fixedRotation, float fixedSpeed=-1)
         {
             MoveTo(target, fixedRotation, fixedSpeed);
